multitasking project - working 20 seconds for 6 milions

+++ core.php
@@ -7,11 +7,20 @@ ini_set('display_startup_errors', 1);
 //error_reporting( E_ALL ^ E_NOTICE );
 error_reporting(E_ALL);
 
-echo ini_get("memory_limit") . "\n";
-ini_set('memory_limit', '2048M');
-echo ini_get("memory_limit") . "\n";
-echo "not real: " . (memory_get_peak_usage(false) / 1024 / 1024) . " MiB\n";
-echo "real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
+$task = $_GET["task"];
+
+if ($task == 0)
+{
+    
+}
+else
+{
+    echo ini_get("memory_limit") . "\n";
+    ini_set('memory_limit', '2048M');
+    echo ini_get("memory_limit") . "\n";
+    echo "not real: " . (memory_get_peak_usage(false) / 1024 / 1024) . " MiB\n";
+    echo "real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
+}
 
 $dbName = "praktykanci";
 $host = "localhost";
@@ -23,6 +32,10 @@ $randomGeneratorTimeLetterGen = 0;
 $randomGeneratorTimeNumber = 0;
 $randomGeneratorTimeSql = 0;
 $sql = "";
+
+$fileCSV = "";
+$startLoop = null;
+$stopLoop = null;
 
 function randomString($length = 15)
 {
@@ -54,12 +67,24 @@ function randomAge()
     return $result;
 }
 
-$dbh = new PDO("pgsql:dbname=$dbName;host=$host", $dbUser, $dbPass);
-$dbh->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-$sqlTableTruncate = "DROP TABLE public.users";
-$dbh->query($sqlTableTruncate);
+$startTimeScript = microtime(true);
+$startTimeGlobal = microtime(true);
 
-$sqlTableCreate = "CREATE TABLE public.users
+//$countRecords = 5000000;
+
+//echo "do zapisania w tabeli: " . number_format($countRecords, 0, ',', ' ') . " rekordów\n";
+
+$startTime = microtime(true);
+
+if ($_GET["task"] == 0)
+{
+    $task = 0;
+    $dbh = new PDO("pgsql:dbname=$dbName;host=$host", $dbUser, $dbPass);
+    $dbh->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+    $sqlTableTruncate = "DROP TABLE public.users";
+    $dbh->query($sqlTableTruncate);
+
+    $sqlTableCreate = "CREATE TABLE public.users
 (
    user_id serial, 
    first_name character varying(15), 
@@ -70,78 +95,114 @@ $sqlTableCreate = "CREATE TABLE public.users
 WITH (
   OIDS = FALSE
 )";
-$sqlTableOwner = 'ALTER TABLE public.users
+    $sqlTableOwner = 'ALTER TABLE public.users
   OWNER TO "user-praktykanci"';
 
-$dbh->query($sqlTableCreate);
-$dbh->query($sqlTableOwner);
-
-$startTimeScript = microtime(true);
-$startTimeGlobal = microtime(true);
-echo "Start: " . date('H:i:s', time()) . " \n";
-
-$countRecords = 5000000;
-
-echo "do zapisania w tabeli: " . number_format($countRecords, 0, ',', ' ') . " rekordów\n";
-
-$startTime = microtime(true);
-
-$fileCSV = "plik.csv";
-
-$output = shell_exec('mkdir /tmp/ram');
-echo $output."\n";
-        
-$output = shell_exec('mount -t tmpfs -o size=512m tmpfs /tmp/ram');
-echo $output."\n";
-
-try
-{
-    $fileHandler = fopen('/tmp/ram/'.$fileCSV, 'w');
-
-    $dataTable = array('user_id', 'first_name', 'last_name', 'user_age',);
-
-    if ($fileHandler != false)
-    {
-
-        echo "Memory used (before) real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
-        fputcsv($fileHandler, $dataTable);
-
-        for ($i = 0; $i < 5000000; $i ++)
-        {
-            fputcsv($fileHandler, array($i + 1, randomString(15), randomString(15), randomAge()));
-        }
-        
-        fclose($fileHandler);
-
-        echo "Memory used (after) real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
-    }
-    else
-    {
-        echo "File open error";
-    }
+    $dbh->query($sqlTableCreate);
+    $dbh->query($sqlTableOwner);
+    echo "Tabela users została wyczysczona!\n\n";
+//    exit();
 }
-catch (Exception $ex)
+elseif ($_GET["task"] == 1)
 {
-    echo "File Error: " . $ex->getMessage();
+    $task = 1;
+    $fileCSV = "plik" . $task . ".csv";
+    $startLoop = 0;
+    $stopLoop = $task * 2000000;
+}
+elseif ($_GET["task"] == 2)
+{
+    $task = 2;
+    $fileCSV = "plik" . $task . ".csv";
+    $startLoop = ($task - 1) * 2000000;
+    $stopLoop = $task * 2000000;
+}
+elseif ($_GET["task"] == 3)
+{
+    $task = 3;
+    $fileCSV = "plik" . $task . ".csv";
+    $startLoop = ($task - 1) * 2000000;
+    $stopLoop = $task * 2000000;
+}
+elseif ($_GET["task"] == 4)
+{
+    $task = 4;
+    $fileCSV = "plik" . $task . ".csv";
+    $startLoop = ($task - 1) * 1250000;
+    $stopLoop = $task * 1250000;
+}
+else
+{
+    echo "BŁĄÐ: parametr task nie ustawiony!";
+    echo "CSV: " . $fileCSV . "\n";
+    echo "startLoop: " . $startLoop . "\n";
+    echo "stopLoop: " . $stopLoop . "\n";
+    exit();
 }
 
-$sqlBulk = "COPY users (user_id, first_name, last_name, user_age)
-    FROM '/tmp/ram/$fileCSV'
-    CSV 
-    HEADER";
+//$output = shell_exec('rm -rf /tmp/ram');
+//echo $output . "\n";
+//
+//$output = shell_exec('mkdir /tmp/ram');
+//echo $output . "\n";
+//
+//$output = shell_exec('mount -t tmpfs -o size=512m tmpfs /tmp/ram');
+//echo $output . "\n";
+//$fileCSV = "plik.csv";
+//$startLoop = 0;
+//$stopLoop = 1000000;
 
-echo "Memory used (before) real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
-
+if ($task == 0)
+{
+    
+}
+else
+{
     try
     {
+        $fileHandler = fopen('/tmp/ram/' . $fileCSV, 'w');
+        $dataTable = array('user_id', 'first_name', 'last_name', 'user_age',);
+        if ($fileHandler != false)
+        {
+            echo "Memory used (before) real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
+//        fputcsv($fileHandler, $dataTable);
+//        fwrite($fileHandler, "user_id,first_name,last_name,user_age\n");
+            for ($i = $startLoop; $i < $stopLoop; $i ++)
+            {
+//            fputcsv($fileHandler, array($i + 1, randomString(15), randomString(15), randomAge()));
+                $string = ($i + 1 . "," . randomString(15) . "," . randomString(15) . "," . randomAge()) . "\n";
+                fwrite($fileHandler, $string);
+            }
+
+            fclose($fileHandler);
+            echo "Memory used (after) real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
+        }
+        else
+        {
+            echo "File open error";
+        }
+    }
+    catch (Exception $ex)
+    {
+        echo "File Error: " . $ex->getMessage();
+    }
+    $sqlBulk = "COPY users (user_id, first_name, last_name, user_age)
+    FROM '/tmp/ram/$fileCSV'
+    DELIMITER ','";
+    echo "Memory used (before) real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
+    try
+    {
+        $dbh = new PDO("pgsql:dbname=$dbName;host=$host", $dbUser, $dbPass);
+        $dbh->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
         $dbh->query($sqlBulk);
     }
     catch (PDOException $e)
     {
         echo 'PDO error: ' . $e->getMessage() . "\n\n";
     }
+    echo "Memory used (after) real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
+}
 
-echo "Memory used (after) real: " . (memory_get_peak_usage(true) / 1024 / 1024) . " MiB\n\n";
 
 $endTime = microtime(true);
 $randomGeneratorTimeSql += ($endTime - $startTime);
@@ -151,17 +212,25 @@ $endTimeScript = microtime(true);
 
 $allTimeScript = ($endTimeScript - $startTimeScript);
 $endTimeGlobal = microtime(true);
-echo "Koniec: " . date('H:i:s', time()) . "\n" . ($endTimeGlobal - $startTimeGlobal) . " s\n";
-echo number_format($allTimeScript, 16, '.', ' ') . " sek - Czas całego skryptu\n";
-echo number_format($randomGeneratorTimeSql, 16, '.', ' ') . " sek - Czas SQL-a\n";
-echo number_format($randomGeneratorTimeLetter, 16, '.', ' ') . " sek - randomString()\n";
-echo number_format($randomGeneratorTimeNumber, 16, '.', ' ') . " sek - randomAge()\n\n";
 
-$allPercentNumber = $allTimeScript;
-$sqlPercent = (($randomGeneratorTimeSql) * 100) / $allPercentNumber;
-$randomPercent = ((($randomGeneratorTimeLetter) + ($randomGeneratorTimeNumber)) * 100) / $allPercentNumber;
-echo "Proporcje czasów działania skryptu\n\n";
-echo number_format($sqlPercent, 4, '.', ' ') . "% - czas SQL-a\n";
-echo " " . number_format($randomPercent, 4, '.', ' ') . "% - czas generowania danych\n";
+if ($task == 0)
+{
+    
+}
+else
+{
+    echo "Koniec: " . date('H:i:s', time()) . "\n" . ($endTimeGlobal - $startTimeGlobal) . " s\n";
+    echo number_format($allTimeScript, 16, '.', ' ') . " sek - Czas całego skryptu\n";
+    echo number_format($randomGeneratorTimeSql, 16, '.', ' ') . " sek - Czas SQL-a\n";
+    echo number_format($randomGeneratorTimeLetter, 16, '.', ' ') . " sek - randomString()\n";
+    echo number_format($randomGeneratorTimeNumber, 16, '.', ' ') . " sek - randomAge()\n\n";
+
+    $allPercentNumber = $allTimeScript;
+    $sqlPercent = (($randomGeneratorTimeSql) * 100) / $allPercentNumber;
+    $randomPercent = ((($randomGeneratorTimeLetter) + ($randomGeneratorTimeNumber)) * 100) / $allPercentNumber;
+    echo "Proporcje czasów działania skryptu\n\n";
+    echo number_format($sqlPercent, 4, '.', ' ') . "% - czas SQL-a\n";
+    echo " " . number_format($randomPercent, 4, '.', ' ') . "% - czas generowania danych\n";
+}
 
 echo "</pre>";
